requirements.txt
streamlit
langchain
openai
PyPDF2
langchain-community
PyYAML
dotenv
streamlit-authenticator
